Move `ActionLog` into its own file
